requirements.txt
aiogram==3.0.0b7
Pillow==10.0.0
allpairspy==2.5.0
openpyxl==3.1.2
python-dotenv==1.0.0